requirements.txt
python-dotenv==1.0.1
reedsolo==1.7.0
setuptools==75.1.0
wheel==0.44.0
